pyproject.toml
[tool.poetry]
name = "email-newsletter"
version = "0.1.3"
description = "AI Newsletter generator for KYC and AML compliance topics"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
pyyaml = "^6.0"
jinja2 = "^3.1.2"
premailer = "^3.10.0"
streamlit = "^1.46.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
render-newsletter = "render_newsletter:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 